refactor: reorganize build script imports into filters directory
